+++ tests/Integrations/PDO/PDOTest.php
@@ -2,6 +2,7 @@
 
 namespace DDTrace\Tests\Integrations\PDO;
 
+use DDTrace\Configuration;
 use DDTrace\Integrations\IntegrationsLoader;
 use DDTrace\Tests\Common\IntegrationTestCase;
 use DDTrace\Tests\Common\SpanAssertion;
@@ -293,6 +294,23 @@ final class PDOTest extends IntegrationTestCase
                 ->setError('PDOException', 'Sql error')
                 ->withExactTags($this->baseTags()),
         ]);
+    }
+
+    public function testLimitedTracerPDO()
+    {
+        $query = "SELECT * FROM tests WHERE id = ?";
+        $traces = $this->isolateLimitedTracer(function () use ($query) {
+            $pdo = $this->pdoInstance();
+            $stmt = $pdo->prepare($query);
+            $stmt->execute([1]);
+            $results = $stmt->fetchAll();
+            $this->assertEquals('Tom', $results[0]['name']);
+            $stmt->closeCursor();
+            $stmt = null;
+            $pdo = null;
+        });
+
+        $this->assertEmpty($traces);
     }
 
     private function pdoInstance()